SpieleMA custom fine calculation in overdue list
the custom fines are 1.5 Euros for the first week
irrespective of the number of overdue items.
And 0.5 Euros per item for every further overdue week.

we still have to check if this works correctly on the day
that the items are due(and exactly one week later)

the calculation has to be added to the circulation as well

custom for SpieleMA; don't merge