accept fines with float values in the fines list
both comma and dot are accepted as decimal separator
I think it would be better to use the decimal separator
of the current language. But I don't know which of the
several possible ways to do that would work correctly
with different operating systems and languages.
Ideally we would have a library which handles it for the
whole project