SpieleMA: improve custom fine calculation
always count the current day as a non-holiday
for the fines. This is necessary, because
we open on tuesday, but it is not an open day
in slims, because no games should be due on
tuesdays and only one day per week
should be counted for the fines.

custom for SpieleMA; don't merge